fix: GetSidSubAuthorityCount returns a pointer, not an int
AccessViolationException on every hotkey press was caused by a wrong
P/Invoke signature in NativeMethods: GetSidSubAuthorityCount was
declared as returning int. The real Win32 signature returns PUCHAR — a
pointer to a single byte holding the count. Treating that pointer as
an integer count then passing `count - 1` into GetSidSubAuthority sent
us to a bogus memory address, hence the AV.

Fix:
- Change GetSidSubAuthorityCount's P/Invoke return type to nint.
- Dereference with Marshal.ReadByte to get the actual count.
- Add null-pointer guards on sid / countPtr / subPtr for defence
  in depth.

Also align the SettingsTests assertion with the (user-chosen) "Win+Z"
default hotkey so the test suite stays green.

+++ src/WhisperTray.App/Adapters/NativeMethods.cs
@@ -76,9 +76,11 @@ internal static class NativeMethods
         uint tokenInformationLength,
         out uint returnLength);
 
+    // Returns PUCHAR (pointer to a single byte) — must be dereferenced.
     [DllImport("advapi32.dll", SetLastError = true)]
-    public static extern int GetSidSubAuthorityCount(nint sid);
+    public static extern nint GetSidSubAuthorityCount(nint sid);
 
+    // Returns PDWORD (pointer to a 32-bit authority value) — must be dereferenced.
     [DllImport("advapi32.dll", SetLastError = true)]
     public static extern nint GetSidSubAuthority(nint sid, uint nSubAuthority);
